requirements.txt
modoboa-admin
reportlab
rfc6266
